pyproject.toml
[project]
name = "private-clinick"
version = "0.1.0"
description = ""
authors = [
    {name = "kolykonon",email = "kononikhinnikolay@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "private_clinick", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
